finish implementing

+++ src/engine-types/engine-types.controller.ts
@@ -8,10 +8,13 @@ import {
   Param,
   Post,
   Put,
+  UseGuards,
 } from '@nestjs/common';
 import { EngineTypesService } from './engine-types.service';
 import { Prisma } from '@prisma/client';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 
+@UseGuards(JwtAuthGuard)
 @Controller('engine-types')
 export class EngineTypesController {
   constructor(private engineTypesService: EngineTypesService) {}
@@ -35,14 +38,18 @@ export class EngineTypesController {
 
   @Put(':id')
   async update(
-    @Param() id: string,
+    @Param('id') id: string,
     @Body() data: Prisma.EngineTypeUpdateInput,
   ) {
+    let engineType = await this.engineTypesService.findOne({ id: +id });
+    if (!engineType) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     return this.engineTypesService.update({ where: { id: +id }, data });
   }
 
   @Delete(':id')
-  async delete(@Param() id: string) {
+  async delete(@Param('id') id: string) {
+    let engineType = await this.engineTypesService.findOne({ id: +id });
+    if (!engineType) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     return this.engineTypesService.delete({ id: +id });
   }
 }
