+++ src/engines/engines.controller.ts
@@ -8,17 +8,20 @@ import {
   Param,
   Post,
   Put,
+  UseGuards,
 } from '@nestjs/common';
 import { EnginesService } from './engines.service';
 import { Prisma } from '@prisma/client';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 
+@UseGuards(JwtAuthGuard)
 @Controller('engines')
 export class EnginesController {
   constructor(private enginesService: EnginesService) {}
 
   @Get(':id')
   async findOne(@Param('id') id: string) {
-    const engine = await this.enginesService.findOne({ id: +id });
+    let engine = await this.enginesService.findOne({ id: +id });
     if (!engine) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     return engine;
   }
@@ -34,12 +37,19 @@ export class EnginesController {
   }
 
   @Put(':id')
-  async update(@Param() id: string, @Body() data: Prisma.EngineUpdateInput) {
+  async update(
+    @Param('id') id: string,
+    @Body() data: Prisma.EngineUpdateInput,
+  ) {
+    let engine = await this.enginesService.findOne({ id: +id });
+    if (!engine) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     return this.enginesService.update({ where: { id: +id }, data });
   }
 
   @Delete(':id')
-  async delete(@Param() id: string) {
+  async delete(@Param('id') id: string) {
+    let engine = await this.enginesService.findOne({ id: +id });
+    if (!engine) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     return this.enginesService.delete({ id: +id });
   }
 }
